+++ Sources/OZCharts/Interaction/ChartViewportState.swift
@@ -9,14 +9,25 @@
 public struct ChartViewportState: Equatable {
     public var visibleXDomain: ClosedRange<Double>?
     public var visibleYDomain: ClosedRange<Double>?
+    public var liveTrackingStatus: ChartLiveTrackingStatus
+    public var command: ChartViewportCommand?
 
     public init(
         visibleXDomain: ClosedRange<Double>? = nil,
-        visibleYDomain: ClosedRange<Double>? = nil
+        visibleYDomain: ClosedRange<Double>? = nil,
+        liveTrackingStatus: ChartLiveTrackingStatus = .inactive,
+        command: ChartViewportCommand? = nil
     ) {
         self.visibleXDomain = visibleXDomain
         self.visibleYDomain = visibleYDomain
+        self.liveTrackingStatus = liveTrackingStatus
+        self.command = command
     }
 
     public static let automatic = ChartViewportState()
+    public static let jumpToLatest = ChartViewportState(command: .jumpToLatest)
+
+    public mutating func requestJumpToLatest() {
+        command = .jumpToLatest
+    }
 }